refactor: type들 types 폴더 안의 파일로 다 옮김
