+++ src/components/Selection/RestaurantSelectionResult.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import RestaurantDetail from '../RestaurantDetail';
-import { FetchedStayDataType } from '../../apis/publicAPI';
+import { FetchedStayDataType } from '../../types/apiDataTypes';
 import noimg from '../../assets/noimg.avif';
 import { useQuery } from 'react-query';
 import { fetchRestaurantData } from '../../apis/publicAPI';
