+++ src/pages/DetailPage/DetailPage.tsx
@@ -1,6 +1,7 @@
 import { useQuery } from 'react-query';
 import { Link, useParams, useNavigate } from 'react-router-dom';
-import { fetchSpotDetailData, FetchedStayDataType } from '../../apis/publicAPI';
+import { fetchSpotDetailData } from '../../apis/publicAPI';
+import { FetchedStayDataType } from '../../types/apiDataTypes';
 import Loader from '../../components/Loader/Loader';
 import { useEffect } from 'react';
 import { doc, setDoc, getDoc, updateDoc, increment } from 'firebase/firestore';
