+++ src/types/apiDataTypes.ts
@@ -1,6 +1,20 @@
+export interface recCnt {
+  [key: string]: string;
+}
+
+export type FetchedStayDataType = Omit<recCnt, 'viewCnt' | 'likeCnt'>;
+
+export interface recCnts extends Array<recCnt> {}
+
 export interface PageDataTypes {
   [key: string]: any;
 }
+
+export interface EachRankType {
+  [key: string]: any;
+}
+
+export interface RankTypeList extends Array<EachRankType> {}
 
 export interface DetailDataTypes {
   // [x: string]: any;
